// src/types/invidious.ts
export interface ThumbnailObject {
  quality: string;
  url: string;
  width: number;
  height: number;
}

export interface VideoObject {
  type: "video" | "shortVideo" | string;
  title: string;
  videoId: string;
  author: string;
  authorId: string;
  authorUrl: string;
  authorVerified?: boolean;
  videoThumbnails: ThumbnailObject[];
  authorThumbnails?: ThumbnailObject[];
  description?: string;
  descriptionHtml?: string;
  viewCount?: number;
  viewCountText?: string;
  lengthSeconds?: number;
  published?: number;
  publishedText?: string;
  premiereTimestamp?: number;
  liveNow?: boolean;
  premium?: boolean;
  paid?: boolean;
  isUpcoming?: boolean;
  isNew?: boolean;
  is4k?: boolean;
  is8k?: boolean;
  isVr180?: boolean;
  isVr360?: boolean;
  is3d?: boolean;
  hasCaptions?: boolean;
  playlistId?: string;
  playlistThumbnail?: string;
}

export interface ChannelObject {
  type: "channel";
  author: string;
  authorId: string;
  authorUrl: string;
  authorVerified?: boolean;
  authorThumbnails: ThumbnailObject[];
  autoGenerated?: boolean;
  subCount?: number;
  videoCount?: number;
  description?: string;
  descriptionHtml?: string;
}

export interface PlaylistVideoObject {
  title: string;
  videoId: string;
  author?: string;
  authorId?: string;
  authorUrl?: string;
  lengthSeconds?: number;
  index?: number;
  indexId?: string;
  videoThumbnails: ThumbnailObject[];
}

export interface PlaylistObject {
  type: "playlist" | "invidiousPlaylist";
  title: string;
  playlistId: string;
  playlistThumbnail?: string;
  author: string;
  authorId?: string | null;
  authorUrl?: string | null;
  authorVerified?: boolean;
  videoCount?: number;
  videos?: PlaylistVideoObject[];
}
// ...
export interface HashtagObject {
  type: "hashtag" | string;
  title?: string;
  hashtag?: string;
  hashtagUrl?: string;
  description?: string;
  descriptionHtml?: string;
}

export type SearchResultObject = VideoObject | ChannelObject | PlaylistObject | HashtagObject;